feat: limit the image size to the device height
In this commit we use the React Native `useWindowDimensions` module to
get the size of the current device window. Any change in device window
size will update the value returned by this function and therefore cause
a re-render in our image component. This allows us to resize the image
on the fly, whether the device is in portrait or landscape mode. Or
even, like with side-by-side iPad apps, if the window is resized.