feat: limit image size by device width
In this commit, we also use the device width to resize the image. We add
the `fit: 'crop'` imgix rendering API parameter to crop the image within
the width and height limits we've set, rather than resize the image. By
cropping the excess image data we shave bytes and decrease load times.
We also avoid distorting the image.