feat: have image resolution match device DPR
In this commit, we use the React Native `PixelRatio` module to get the
current device's device pixel ratio (DPR). We then pass that value as an
imgix `dpr` param. This ensures that the image rendered will always
match the current device's resolution. Not only does this save bytes
when loading images on low resolution devices, but it also future proofs
your React Native app so that it can match any future resolution
increases.