feat: add lqip to render first
In this commit, we create a low quality image placeholder (LQIP). To do
this, we use the imgix rendering API to create a low resolution version
of our image that we render first. Then we render the full resolution
image on top of the low resolution image. This is a common technique to
show something to the user while the full image loads.

1. Create a low resolution, blurred, image URL
2. Add a new View layer
3. Add our existing image component to the new view layer
4. Add another image component to the view layer and set the low
   resolution image as its source.
5. Style the images so that they are absolutely positioned inside the
   view layer, allowing them to render on top of the last rendered
image.